fixes and clean up
