+++ src/app/sitemap.ts
@@ -1,15 +1,15 @@
 import type { MetadataRoute } from "next";
 import { siteUrl } from "@/shared/lib/site-metadata";
 
-const publicRoutes = ["", "/sign-in", "/sign-up"];
+const marketingRoutes = [""]; // Only expose the canonical landing page
 
 export default function sitemap(): MetadataRoute.Sitemap {
   const lastModified = new Date();
 
-  return publicRoutes.map((path) => ({
+  return marketingRoutes.map((path) => ({
     url: `${siteUrl}${path}`,
     lastModified,
     changeFrequency: "monthly",
-    priority: path === "" ? 1 : 0.6,
+    priority: 1,
   }));
 }
